benchmark/requirements_benchmark.txt
# Requirements for PFAS-atlas benchmark
pandas>=1.3.0
numpy>=1.20.0
rdkit-pypi>=2022.3.0
mhfp>=1.9.6
torch>=1.12.0
scikit-learn>=1.0.0